Closes #1375: Migrate Variables from figma_import to dc_bundle
Move variable declarations into variable.rs. This includes creating some structs that are basically copies of their counterpart in figma_schema.

Add a variable_utils.rs file that includes functions to create variables from figma_schema data.
