+++ crates/dc_bundle/src/legacy_definition/element.rs
@@ -19,3 +19,4 @@ pub mod font;
 pub mod geometry;
 pub mod node;
 pub mod path;
+pub mod variable;
