+++ crates/figma_import/src/reflection.rs
@@ -165,33 +165,35 @@ pub fn registry() -> serde_reflection::Result<serde_reflection::Registry> {
     tracer
         .trace_type::<crate::toolkit_schema::StrokeCap>(&samples)
         .expect("couldn't trace StrokeCap");
-    tracer.trace_type::<crate::toolkit_schema::Mode>(&samples).expect("couldn't trace Mode");
     tracer
-        .trace_type::<crate::toolkit_schema::Collection>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::element::variable::Mode>(&samples)
+        .expect("couldn't trace Mode");
+    tracer
+        .trace_type::<dc_bundle::legacy_definition::element::variable::Collection>(&samples)
         .expect("couldn't trace Collection");
     tracer
-        .trace_type::<crate::figma_schema::VariableType>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::element::variable::VariableType>(&samples)
         .expect("couldn't trace VariableType");
     tracer
         .trace_type::<crate::figma_schema::FigmaColor>(&samples)
         .expect("couldn't trace FigmaColor");
     tracer
-        .trace_type::<crate::figma_schema::VariableAlias>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::element::variable::VariableAlias>(&samples)
         .expect("couldn't trace VariableAlias");
     tracer
-        .trace_type::<crate::toolkit_schema::NumOrVar>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::element::variable::NumOrVar>(&samples)
         .expect("couldn't trace NumOrVar");
     tracer
-        .trace_type::<crate::toolkit_schema::ColorOrVar>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::element::variable::ColorOrVar>(&samples)
         .expect("couldn't trace ColorOrVar");
     tracer
-        .trace_type::<crate::toolkit_schema::VariableValue>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::element::variable::VariableValue>(&samples)
         .expect("couldn't trace VariableValue");
     tracer
-        .trace_type::<crate::toolkit_schema::Variable>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::element::variable::Variable>(&samples)
         .expect("couldn't trace Variable");
     tracer
-        .trace_type::<crate::toolkit_schema::VariableMap>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::element::variable::VariableMap>(&samples)
         .expect("couldn't trace VariableMap");
     tracer
         .trace_type::<crate::toolkit_schema::ViewShape>(&samples)
