+++ crates/figma_import/src/toolkit_style.rs
@@ -20,6 +20,7 @@ use dc_bundle::legacy_definition::element::color::Color;
 use dc_bundle::legacy_definition::element::font::{FontFeature, FontStretch, FontStyle};
 use dc_bundle::legacy_definition::element::geometry::Size;
 use dc_bundle::legacy_definition::element::path::{LineHeight, StrokeAlign, StrokeWeight};
+use dc_bundle::legacy_definition::element::variable::{ColorOrVar, NumOrVar};
 use dc_bundle::legacy_definition::layout::layout_style::LayoutStyle;
 use dc_bundle::legacy_definition::modifier::blend::BlendMode;
 use dc_bundle::legacy_definition::modifier::filter::FilterOp;
@@ -30,7 +31,6 @@ use std::collections::HashMap;
 use crate::{
     toolkit_font_style::{FontWeight, TextDecoration},
     toolkit_layout_style::{Display, LayoutSizing, Number, Overflow},
-    toolkit_schema::{ColorOrVar, NumOrVar},
 };
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize, Default)]
